feat: move delta to new structure

+++ tests/test_delta.py
@@ -21,6 +21,7 @@ table = {
     "name": "mytesttable",
     "schema": pa.schema([pa.field("col", pa.string()), pa.field("value", pa.int32())]),
     "partition": "col",
+    "location": "fs",
 }
 
 path = "/tmp/delta"
@@ -29,36 +30,18 @@ path = "/tmp/delta"
 class TestDelta(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        cls.lake = Delta(path)
-
-    @classmethod
-    def tearDownClass(cls):
-        cls.lake.delete_table(table["name"])
-        shutil.rmtree(path)
+        if os.path.exists(path):
+            shutil.rmtree(path)
+        cls.lake = Delta(table, path)
 
     def test_create_table(self):
-        self.lake.add_table(table)
-
         metadata = os.listdir(os.path.join(path, table["name"], "_delta_log"))
         self.assertGreater(len(metadata), 0)
 
-        result = self.lake.add_table(table)
-        self.assertEqual(result, False)
-
     def test_write_table(self):
-        tbl = self.lake.write(table["name"], df, mode="append")
+        tbl = self.lake.append(df)
         self.assertEqual(len(tbl.to_pandas()), 5)
-
-    def test_get_table(self):
-        tbl = self.lake.get_table("asdasd")
-        self.assertEqual(tbl, None)
-
-        tbl = self.lake.write("asdasd", df)
-        self.assertEqual(tbl, None)
-
-    def test_fake_delete(self):
-        tbl = self.lake.delete_table("asdasd")
-        self.assertEqual(tbl, False)
+        self.assertEqual(len(self.lake.read()), 5)
 
 
 if __name__ == "__main__":
